Experimental file system changes
It seems that some titles will attempt to write files to the utility drive
without creating the necessary parent folders. I'm not sure if this is
intended behavior, but whenever a file creation attempt fails because its
parent folder does not exist (and if the file was supposed to be created),
the emulator will create the necessary parent folders first and try the
operation again. It would be nice to confirm this behavior on a real Xbox.

Other changes include the initial implementation of FileInformation
conversions. Combined with the above, Panzer Dragoon ORTA boots up and
even gets in game for a bit before crashing hard due to problems with D3D
emulation (corrupted textures, vertex shaders and more).