Aliassed D3DTexture_LockRect better
D3DTexture_LockRect version 5849 already appeared in 5788, which is
identical to 5233 (the version in between, 5558, is different, by one
OV-pair, so I kept that).

PS: I found this purely by accident, when taking a quick look at jarupxx
past few commits, of which 2c6dfd5b stood out because of the "Removed"
description. I fear more of these little mistakes might have slipped in
with that batch...