Fixed memory leak of pCachedHostVertexBuffer
Now that symbols are renamed, the bug was much easier to find and fix:
CxbxVertexBufferConverter::Restore should always ReleasePatchedStream,
but that function should only free pCachedHostVertexStreamZeroData if it
isn't cached. The pCachedHostVertexBuffer however should always be
Released().
(This bug was spotted looking at apitrace dumps of X-Marles, which
showed a per-frame increasing reference count of vertex buffers.)