VertexBuffer cache now checks a few samples, to avoid hashing the entire buffer
This is so cheap that it can be done every time (while the hashing is
postponed ever longer after each succesfull check).

With this, textures show up correctly again.

Somehow, this change (or one of the last few) seems to run a lot slower.
At least for me, Turok doesn't react as quickly anymore??