FindContainingTypedMemoryBlock : avoid returning the wrong block.
This issue only occurs when FindContainingTypedMemoryBlock is called with an address previously freeed - a so-called 'double-free' (which in itself, is a bug in our code, but still).

Without this additional check, when a double-free occurs, MemoryManager::Free crashes with : "Fatal: MemoryManager attempted to free only a part of a block"
With this additional check, a double-free will crash on : "Fatal: Attempted to free memory that was not allocated via MemoryManager"
Since it was caused by a double-free, the latter is more accurate than the former.