Cache partitions now work as expected, with Format/Mount functions unpatched.
Note: This requires all PartitionX.bin files within %AppData%\Cxbx-Reloaded\EmuDisk\ to be deleted before this branch is executed. After this, it will work as expected without the need to do this again in the future.

There may be a way to auto-detect if we need to do this, and manually update the partition table... Will update if possible.