Unpatch Texture Lock functions and introduce EmuUpdateActiveTextureStages
This fixed an issue in Crash Bandicoot: The Wrath of Cortex in which the
splash screen was corrupt (new issue to this branch)

Currently, this function does no hashing/dirty flags/any checks for
modification, so titles that modify textures on the fly will not display
correctly.

This will be addressed in a later commit.