DO NOT TEST THIS (it probably won't compile).
Ugly hack to get uber shader compiled in, and some initial debugging, tweaking and fixing to get it working as intended. Still a long way to go before it will actually work, but with these few changes, it looks like the correct values are being interpreted by the shader, as it responds to changes according to expectations. Since texture-fetch and vertex-shader inputs don't work yet, there's not much to see yet.